+++ src/main/java/one/dfy/bily/api/user/dto/UserActivity.java
@@ -3,6 +3,7 @@ package one.dfy.bily.api.user.dto;
 import one.dfy.bily.api.inquiry.dto.InquiryPreferredDateInfo;
 import one.dfy.bily.api.reservation.dto.ReservationPreferredDateInfo;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -11,7 +12,7 @@ public record UserActivity(
         String type,
         String spaceName,
         String location,
-        int areaM2,
+        BigDecimal areaM2,
         int areaPy,
         int maxCapacity,
         List<InquiryPreferredDateInfo> inquiryPreferredDateList,
